HSEARCH-4328 Refactor JSR-352 ITs for safer startup/shutdown
