+++ integrationtest/mapper/orm-batch-jsr352/src/test/java/org/hibernate/search/integrationtest/batch/jsr352/massindexing/EntityManagerFactoryRetrievalIT.java
@@ -6,21 +6,39 @@
  */
 package org.hibernate.search.integrationtest.batch.jsr352.massindexing;
 
+import static org.hibernate.search.integrationtest.batch.jsr352.util.JobTestUtil.JOB_TIMEOUT_MS;
+import static org.hibernate.search.util.impl.integrationtest.mapper.orm.OrmUtils.with;
 import static org.junit.Assert.assertEquals;
 
+import java.util.ArrayList;
 import java.util.List;
+import javax.batch.operations.JobOperator;
 import javax.batch.runtime.JobExecution;
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
 
 import org.hibernate.search.batch.jsr352.core.massindexing.MassIndexingJob;
+import org.hibernate.search.integrationtest.batch.jsr352.massindexing.entity.Person;
+import org.hibernate.search.integrationtest.batch.jsr352.massindexing.entity.WhoAmI;
 import org.hibernate.search.integrationtest.batch.jsr352.util.JobTestUtil;
 import org.hibernate.search.integrationtest.batch.jsr352.massindexing.entity.Company;
+import org.hibernate.search.integrationtest.batch.jsr352.util.PersistenceUnitTestUtil;
 
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
 /**
  * @author Mincong Huang
  */
-public class EntityManagerFactoryRetrievalIT extends AbstractBatchIndexingIT {
+public class EntityManagerFactoryRetrievalIT {
+
+	private static final String PERSISTENCE_UNIT_NAME = PersistenceUnitTestUtil.getPersistenceUnitName();
+
+	protected static final int INSTANCES_PER_DATA_TEMPLATE = 100;
+
+	// We have three data templates per entity type (see setup)
+	protected static final int INSTANCE_PER_ENTITY_TYPE = INSTANCES_PER_DATA_TEMPLATE * 3;
 
 	/*
 	 * Make sure to have more than one checkpoint,
@@ -29,6 +47,49 @@ public class EntityManagerFactoryRetrievalIT extends AbstractBatchIndexingIT {
 	private static final int CHECKPOINT_INTERVAL = 10;
 
 	private static final String SESSION_FACTORY_NAME = "primary_session_factory";
+
+	protected JobOperator jobOperator;
+	protected EntityManagerFactory emf;
+
+	@Before
+	public void setup() {
+		jobOperator = JobTestUtil.getAndCheckRuntime();
+		List<Company> companies = new ArrayList<>();
+		List<Person> people = new ArrayList<>();
+		List<WhoAmI> whos = new ArrayList<>();
+		for ( int i = 0; i < INSTANCE_PER_ENTITY_TYPE; i += 3 ) {
+			int index1 = i;
+			int index2 = i + 1;
+			int index3 = i + 2;
+			companies.add( new Company( "Google " + index1 ) );
+			companies.add( new Company( "Red Hat " + index2 ) );
+			companies.add( new Company( "Microsoft " + index3 ) );
+			people.add( new Person( "BG " + index1, "Bill", "Gates" ) );
+			people.add( new Person( "LT " + index2, "Linus", "Torvalds" ) );
+			people.add( new Person( "SJ " + index3, "Steven", "Jobs" ) );
+			whos.add( new WhoAmI( "cid01 " + index1, "id01 " + index1, "uid01 " + index1 ) );
+			whos.add( new WhoAmI( "cid02 " + index2, "id02 " + index2, "uid02 " + index2 ) );
+			whos.add( new WhoAmI( "cid03 " + index3, "id03 " + index3, "uid03 " + index3 ) );
+		}
+
+		emf = Persistence.createEntityManagerFactory( getPersistenceUnitName() );
+		with( emf ).runInTransaction( em -> {
+			companies.forEach( em::persist );
+			people.forEach( em::persist );
+			whos.forEach( em::persist );
+		} );
+	}
+
+	@After
+	public void shutdown() {
+		if ( emf != null ) {
+			emf.close();
+		}
+	}
+
+	protected String getPersistenceUnitName() {
+		return PERSISTENCE_UNIT_NAME;
+	}
 
 	@Test
 	public void defaultNamespace() throws Exception {
