+++ integrationtest/mapper/orm-batch-jsr352/src/test/java/org/hibernate/search/integrationtest/batch/jsr352/util/BackendConfigurations.java
@@ -1,0 +1,31 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.integrationtest.batch.jsr352.util;
+
+import static org.hibernate.search.util.impl.integrationtest.common.rule.BackendConfiguration.BACKEND_TYPE;
+
+import org.hibernate.search.util.impl.integrationtest.backend.elasticsearch.ElasticsearchBackendConfiguration;
+import org.hibernate.search.util.impl.integrationtest.backend.lucene.LuceneBackendConfiguration;
+import org.hibernate.search.util.impl.integrationtest.common.rule.BackendConfiguration;
+
+public final class BackendConfigurations {
+
+	private BackendConfigurations() {
+	}
+
+	public static BackendConfiguration simple() {
+		switch ( BACKEND_TYPE ) {
+			case "lucene":
+				return new LuceneBackendConfiguration();
+			case "elasticsearch":
+				return new ElasticsearchBackendConfiguration();
+			default:
+				throw new IllegalStateException( "Unknown backend type: " + BACKEND_TYPE );
+		}
+	}
+
+}
